feat: add download and upload scripts for instagram highlights
We first use instaloader and python to download all highlights from an account to the download folder. Then we run upload.ts which takes all folders under the download folder and takes all jpgs and mp4s and uploads them to cloudinary. This script also creates a json file which saves meta data of the media. We can then just plug this json into the public folder of the webapp and reference the urls in it.