feat(travel): finalize CountryModal design (for now...)
I experimented with a lot of different versions. I feel like this version is the cleanest. The image is the main focus, we don't waste any space unnecessarily. The main issue is the contrast between the country name and the image which was solved for now using a text border (drop shadow)